requirements.txt
# OCR and PDF processing
ocrmypdf==14.4.0
pdfplumber==0.10.2

# Natural Language Processing
nltk==3.8.1

# Machine Learning and Text Generation
transformers==4.30.2
torch==2.0.1

# Data processing and analysis
numpy==1.24.3
pandas==2.0.2

# YAML parsing (for configuration files)
PyYAML==6.0

# Progress bars 
tqdm==4.65.0

# Jupyter Notebook 
jupyter==1.0.0
